week2/chatbot/requirements.txt
python-dotenv
langchain
langchain-community
langgraph
duckduckgo-search
langchain-groq
psycopg2-binary
langfuse 
langfuse-langchain
